fix: remove polyfills which are out of range of our engines
The polyfills removed are:
- file-url-to-path.js
- mkdir.js
