+++ lib/index.js
@@ -2,7 +2,7 @@ module.exports = {
   ...require('./fs.js'),
   copyFile: require('./copy-file.js'),
   cp: require('./cp/index.js'),
-  mkdir: require('./mkdir/index.js'),
+  mkdir: require('./mkdir.js'),
   mkdtemp: require('./mkdtemp.js'),
   rm: require('./rm/index.js'),
   withTempDir: require('./with-temp-dir.js'),
